Currency Converter & Gas Fee Reformat

+++ TerraSharp.Rest.Legacy/Json/Tx/Transaction/TxResponse.cs
@@ -12,8 +12,8 @@ namespace Terra.Microsoft.Rest.Tx.Transaction
         public string raw_log { get; set; }
         public TxInfoLogsJSON[] logs { get; set; }
         public string info { get; set; }
-        public long gas_wanted { get; set; }
-        public long gas_used { get; set; }
+        public double gas_wanted { get; set; }
+        public double gas_used { get; set; }
         public TxValueJSON tx { get; set; }
         public DateTime timestamp { get; set; }
         public TxInfoEvent[] events { get; set; }
